requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.31
python-dotenv==1.0.0
pydantic==2.8.0
httpx==0.25.1
redis==5.0.1
celery==5.3.4
python-multipart==0.0.19
requests~=2.32.4